Upgrade to integrate with version 2 of the API
The change upgrades the front end so that it integrates with the latest version (V2) of the middleware API. There is now a new data format for the data returned by the API, where documents are returned in a single list rather than being grouped below a single permit. Also, upload date formatting is different and file sizes are now returned in byes rather than MB, so the front end needed to be changed to preserve the on-screen formatting of these data.

There is also a modified check to the back end to check if a permit exists - this is now accomplished using a HTTP HEAD method in the Enter Permit Details screen.

Also, the transition from the Enter Permit Details screen to the View Permit Details screen has been reworked so that it now uses a deep link rather than query parameter, thereby providing deep link ability whilst also avoiding duplication in unit tests.