Permit number reformatting for exact match search
Story 19604

This change ensures that exact match searches are carried out when a user queries for documents via the Enter Permit Number page.

The selected register is mapped to the expected format contained in the metadata. The entered permit number is sanitised and formatted using some register-specific logic.

Note that the logic for register 'Water Quality Discharge Consents' has yet to be defined so this matching logic will need to be added at a later date.