Search across permits
Story 7174

This change adds the ability to search for documents regardless of the permit information. This is achieved my modifying the existing view-permit-documents route so that it searches using a search term rather than a permit number.

The concept of 'page mode' is introduced. If the page mode is 'search' then the search is carried out using a search term containing one or more words to search for in the document metadata.

The following metadata fields are searched: 'DocTitle', 'CustomerOperatorName', 'SiteName', 'FacilityAddressPostcode'

The Contact page has been modified as there is no longer any permit data to display back to the user or include in emails when they have searched via this new search facility rather than by permit number.